refactor(inventory): share payment method labels across client and backend
